Refactor JS, remove Browserify, simplify JS for components
