+++ package.json
@@ -18,7 +18,6 @@
     "@frctl/mandelbrot": "^1.1.0",
     "@frctl/twig": "^1.0.0-beta.2",
     "autoprefixer": "^6.4.1",
-    "browserify": "^14.3.0",
     "combobo": "1.0.2",
     "del": "^2.2.2",
     "eslint": "^3.13.1",
@@ -28,6 +27,10 @@
     "gulp-imagemin": "^3.1.1",
     "gulp-postcss": "^6.1.1",
     "gulp-sass": "^2.3.2",
-    "vinyl-source-stream": "^1.1.0"
+    "npm": "^5.7.1"
+  },
+  "devDependencies": {
+    "gulp-concat": "^2.6.1",
+    "gulp-uglify": "^3.0.0"
   }
 }
