+++ src/@koop-components/_previews/preview.default.handlebars
@@ -9,7 +9,6 @@
 </head>
 <body class="preview">
   {{{ yield }}}
-  <script src="{{~ path '/js/polyfills.js' ~}}"></script>
   <script src="{{~ path '/js/main.js' ~}}"></script>
 </body>
 </html>
