--- src/@koop-components/common/collapsible/decorators/initCollapsible.js
@@ -1,13 +1,0 @@
-var dom = require( 'helpers/dom' );
-var collapsibles = require( '../collapsible' );
-
-module.exports = function initCollapsible( element ) {
-  var showInitially = dom.$('.collapsible--initially-visible', element ).length > 0;
-
-  if ( showInitially === true ){
-    collapsibles.show( element );
-  }
-  else {
-    collapsibles.hide( element );
-  }
-};
