--- src/@koop-components/common/modal/handlers/openModal.js
@@ -1,6 +1,0 @@
-module.exports = function openModal( element ) {
-  var modal = require( 'core/modal' );
-  var modalElement = document.getElementById( element.getAttribute( 'data-modal' ) );
-
-  modal.open( modalElement );
-};
