--- src/@koop-components/common/tabs/handlers/openPanel.js
@@ -1,7 +1,0 @@
-module.exports = function openTab( element, event ) {
-  var tabs = require( 'common/tabs/tabs' );
-
-  event.preventDefault();
-
-  tabs.openPanel( element );
-};
