--- src/@koop-components/page-components/search/searchToggle.js
@@ -1,22 +1,0 @@
-var searchToggle = {
-  'context': {
-    'searchTermHiddenClass' : 'search__term--hidden',
-    'searchTermAnimateClass' : 'search__term--animating',
-    'open' : 'Open zoekveld',
-    'close' : 'Sluit zoekveld',
-    'submit' : 'Zoek'
-  },
-  'handleInputChange': function( event ) {
-    var input = event.target;
-    var button = input.form.querySelector( 'button' );
-
-    if ( input.value.length > 0 ) {
-      button.textContent = searchToggle.context.submit;
-    }
-    else {
-      button.textContent = searchToggle.context.close;
-    }
-  }
-};
-
-module.exports = searchToggle;
