--- src/assets/js/decorators/addMobileFoldability.js
@@ -1,28 +1,0 @@
-var toggle = require( 'handlers/toggleSidebar' );
-
-module.exports = function addMobileFoldability( element ) {
-  var button = document.createElement( 'button' );
-  var labels = {
-    open: 'Open sidebar',
-    close: 'Sluit sidebar'
-  };
-
-  // set data to button
-  button.classList.add( 'hidden-desktop' );
-  button.type = 'button';
-  button.setAttribute( 'data-handler', 'toggle-sidebar' );
-  button.setAttribute( 'aria-controls', element.id );
-  button.setAttribute( 'data-toggle-open', labels.open );
-  button.setAttribute( 'data-toggle-close', labels.close );
-
-  // set initial state
-  button.setAttribute( 'aria-expanded', 'true' );
-  button.textContent = labels.close;
-
-  element.before( button );
-
-  // apply first time
-  if ( !( window.matchMedia( '(min-width: 50em)' ).matches ) ) {
-    toggle( button );
-  }
-};
