+++ src/assets/js/decorators/hide-self.js
@@ -1,0 +1,5 @@
+ohnl.decorate({
+  'hide-self': function(element) {
+    ohnl.ui.hide( element );
+  }
+});
