--- src/assets/js/decorators/hideSelf.js
@@ -1,5 +1,0 @@
-var ui = require( '../helpers/ui' );
-
-module.exports = function hideSelf( element ) {
-  ui.hide( element );
-};
